Create test cases for remaining tests
